adjust registration
  * remove form reset button
  * restructure information about the course into a description list
  * restructure the form to use textareas
  * communication with the backend is done in JavaScript using AJAX
  * frontend depends on jQuery, but we only load the library on pages
    that require it, viz. registration.html.

The registration page is a new type of page, so-called "verbatim".
Why do we need this? Normally, content from the pages/ directory is
compiled using Pandoc. This is great for markdown files, but for HTML
content, it strips out all our custom classes, which are necessary for a
good-looking registration page.
Verbatim pages are not compiled with pandoc, but instead simply by
reading their contents. This is different from copyFileCompiler which
does not allow using templates!