backend: remove unnecessary fields and improve mailer
Now an SMTP connection is established whenever we need to send an
email. The previous strategy was to open a connection at the startup
of the backend, but nothing was done to ensure that the connection
remained alive, which is unsafe. Instead, by creating a new connection
each time, we ensure that the connection isn't closed!